requirements.txt
Django==1.11.5
django-sslserver==0.20
decorator==4.1.2
mohawk==0.3.4
pyjwkest==1.3.6
cryptography==2.0.3
PyBrowserID==0.12.0
pycryptodome==3.4.7
django-cors-headers==2.1.0
whitenoise==3.3.0
